d02-orm/requirements.txt
Django==5.2.9
sqlparse==0.5.4
psycopg2-binary==2.9.6